Move getTestMethodRangesFromDocument to alFileHelper
